refactor more message related types
