+++ ts/a11y/speech/MessageTypes.ts
@@ -1,4 +1,40 @@
+/*************************************************************
+ *
+ *  Copyright (c) 2018-2025 The MathJax Consortium
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+/**
+ * @file  Types for Web worker communication.
+ *
+ * @author v.sorge@mathjax.org (Volker Sorge)
+ */
+
 export interface PromiseFunctions {
   resolve: () => void;
   reject: (cmd: string) => void;
 }
+
+export type Message = { [key: string]: any };
+
+export type WorkerCommand = {
+  cmd: string;
+  debug: boolean;
+  data: Message;
+};
+
+export type PoolCommand = {
+  cmd: string;
+  data: WorkerCommand | Message;
+};
